# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Service Base FastAPI",
    "description": "Version 1 of the FastAPI Application",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/app/v1" }],
  "paths": {
...
    "/user-api-keys": {
      "get": {
        "tags": ["User API Keys"],
        "summary": "User Api Keys:List User Api Keys",
        "description": "Retrieve a list of user API keys with filtering, sorting, and pagination.",
        "operationId": "user_api_keys_list_user_api_keys_user_api_keys_get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sorted_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "default": "id",
              "title": "Sorted By"
            }
          },
          {
            "name": "sorted_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sorted Order"
            }
          },
          {
            "name": "name__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Name  Icontains"
            }
          },
          {
            "name": "api_key__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Api Key  Icontains"
            }
          },
          {
            "name": "expires_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Expires At  Gte"
            }
          },
          {
            "name": "expires_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Expires At  Lte"
            }
          },
          {
            "name": "allowed_origin__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Allowed Origin  Icontains"
            }
          },
          {
            "name": "allowed_ip__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Allowed Ip  Icontains"
            }
          },
          {
            "name": "created_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Gte"
            }
          },
          {
            "name": "created_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Lte"
            }
          },
          {
            "name": "updated_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Gte"
            }
          },
          {
            "name": "updated_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Lte"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserApiKeyListRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["User API Keys"],
        "summary": "User Api Keys:Create User Api Key",
        "description": "Create a new user API key.",
        "operationId": "user_api_keys_create_user_api_key_user_api_keys_post",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserApiKeyCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserApiKeyRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/user-api-keys/{user_api_key_id}": {
      "get": {
        "tags": ["User API Keys"],
        "summary": "User Api Keys:Get User Api Keys",
        "description": "Retrieve a user API key by its ID.",
        "operationId": "user_api_keys_get_user_api_keys_user_api_keys__user_api_key_id__get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "user_api_key_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Api Key Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserApiKeyRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["User API Keys"],
        "summary": "User Api Keys:Update User Api Key",
        "description": "Update an existing user API key.",
        "operationId": "user_api_keys_update_user_api_key_user_api_keys__user_api_key_id__patch",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "user_api_key_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Api Key Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserApiKeyUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserApiKeyRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["User API Keys"],
        "summary": "User Api Keys:Delete User Api Key",
        "description": "Delete a user API key.",
        "operationId": "user_api_keys_delete_user_api_key_user_api_keys__user_api_key_id__delete",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "user_api_key_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "User Api Key Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/user-api-keys/{user_api_key_id}",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/organizations/{organization_id}/users": {
      "get": {
        "tags": ["Organization User"],
        "summary": "Organizations:Users",
        "operationId": "organizations_users_organizations__organization_id__users_get",
        "security": [{ "OAuth2PasswordBearer": [] }, { "APIKeyCookie": [] }],
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Organization Id" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "sorted_by",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "default": "id",
              "title": "Sorted By"
            }
          },
          {
            "name": "sorted_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Sorted Order"
            }
          },
          {
            "name": "email__icontains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [{ "type": "string" }, { "type": "null" }],
              "title": "Email  Icontains"
            }
          },
          {
            "name": "created_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Gte"
            }
          },
          {
            "name": "created_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Created At  Lte"
            }
          },
          {
            "name": "updated_at__gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Gte"
            }
          },
          {
            "name": "updated_at__lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date-time" },
                { "type": "null" }
              ],
              "title": "Updated At  Lte"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListRead" }
              }
            }
          },
          "422": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "examples": {
                  "validation_error": {
                    "summary": "Validation failed on input data.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unprocessable Entity",
                      "status": 422,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users",
                      "errors": [
                        {
                          "status": "422",
                          "code": "validation_error",
                          "title": "Validation Error",
                          "detail": "The input data did not pass validation checks.",
                          "source": { "pointer": "#/body" }
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": {
            "description": "Unauthorized access.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Authentication credentials missing or invalid.",
                    "value": {
                      "type": "about:blank",
                      "title": "Unauthorized",
                      "status": 401,
                      "instance": "http://127.0.0.1:8000/app/v1/organizations/{organization_id}/users",
                      "errors": [
                        {
                          "status": "401",
                          "code": "unauthorized",
                          "title": "Unauthorized",
                          "detail": "Authentication credentials were not provided or are invalid."
                        }
                      ]
                    }
                  }
                },
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "about:blank"
          },
          "title": { "type": "string", "title": "Title" },
          "status": { "type": "integer", "title": "Status" },
          "instance": { "type": "string", "title": "Instance" },
          "errors": {
            "items": { "$ref": "#/components/schemas/ErrorDetail" },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": ["title", "status", "instance", "errors"],
        "title": "ErrorResponse"
      },
...
      "ListResponseMeta": {
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Limit",
            "description": "Maximum number of items to retrieve",
            "default": 100
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset",
            "description": "Starting position for retrieval",
            "default": 0
          },
          "sorted_by": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Sorted By",
            "description": "Field name to sort by",
            "default": "id"
          },
          "sorted_order": {
            "type": "string",
            "title": "Sorted Order",
            "description": "Sort order: asc or desc",
            "default": "asc"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "description": "Total number of items matching the search criteria.",
            "example": 100
          }
        },
        "type": "object",
        "required": ["total_count"],
        "title": "ListResponseMeta"
      },
...
      "UserApiKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Name of the API key",
            "example": "My API Key"
          },
          "expires_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Expires At",
            "description": "Expiration datetime of the API key",
            "example": "2024-12-31T23:59:59Z"
          },
          "allowed_origin": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Origin",
            "description": "CORS allowed origin",
            "example": "https://example.com"
          },
          "allowed_ip": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Ip",
            "description": "Allowed IP address",
            "example": "192.168.1.1"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "UserApiKeyCreate"
      },
      "UserApiKeyListRead": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/ListResponseMeta" },
          "data": {
            "items": { "$ref": "#/components/schemas/ApiKeyRead" },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["meta", "data"],
        "title": "UserApiKeyListRead"
      },
      "UserApiKeyRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "The ID of the object",
            "example": "abcd1234xyzc"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Name of the API key",
            "example": "My API Key"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "The actual API key",
            "example": "1234567890abcdef"
          },
          "expires_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Expires At",
            "description": "Expiration datetime of the API key",
            "example": "2024-12-31T23:59:59Z"
          },
          "allowed_origin": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Origin",
            "description": "CORS allowed origin",
            "example": "https://example.com"
          },
          "allowed_ip": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Ip",
            "description": "Allowed IP address",
            "example": "192.168.1.1"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation datetime of the API key",
            "example": "2023-10-01T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last updated datetime of the API key",
            "example": "2023-10-01T12:00:00Z"
          }
        },
        "type": "object",
        "required": ["id", "name", "api_key", "created_at", "updated_at"],
        "title": "UserApiKeyRead"
      },
      "UserApiKeyUpdate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Name of the API key",
            "example": "My API Key"
          },
          "expires_at": {
            "anyOf": [
              { "type": "string", "format": "date-time" },
              { "type": "null" }
            ],
            "title": "Expires At",
            "description": "Expiration datetime of the API key",
            "example": "2024-12-31T23:59:59Z"
          },
          "allowed_origin": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Origin",
            "description": "CORS allowed origin",
            "example": "https://example.com"
          },
          "allowed_ip": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Allowed Ip",
            "description": "Allowed IP address",
            "example": "192.168.1.1"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "UserApiKeyUpdate"
      },
...
      "UserListRead": {
        "properties": {
          "meta": { "$ref": "#/components/schemas/ListResponseMeta" },
          "data": {
            "items": { "$ref": "#/components/schemas/UserRead" },
            "type": "array",
            "title": "Data"
          }
        },
        "type": "object",
        "required": ["meta", "data"],
        "title": "UserListRead"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "auth/jwt/login" } }
      },
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}